[JE2-252] Refactor code reports
