+++ frontend/src/features/reports/utils/taskReportConstants.ts
@@ -1,7 +1,3 @@
-export const middleWidth = 530
-export const coefficient = 0.9
-export const minWidth = 350
-export const maxWidth = 600
 export const mdWidth = 768
 export const minHeight = 300
 export const maxHeight = 400
